implement INIT_SPACE for Contributor

+++ programs/aixblock/src/constants.rs
@@ -1,4 +1,4 @@
-pub const ANCHOR_DISCRIMINATOR: u8 = 8;
+pub const ANCHOR_DISCRIMINATOR: usize = 8;
 
 pub const SEED_MINT_ACCOUNT: &[u8] = b"mint";
 pub const MINT_DECIMALS: u8 = 6;
